Перенес connect из <GenreFilterItem /> полностью в <GenreFilterList />
